sending out email to talk2us at provisional approval of a claim assessment considerd a total loss

+++ app/Http/Controllers/HeadAssessorController.php
@@ -405,6 +405,7 @@ class HeadAssessorController extends Controller
                 $thirdPartyDriver = isset($request->thirdPartyDriver) ? $request->thirdPartyDriver : null;
                 $thirdPartyPolicy = isset($request->thirdPartyPolicy) ? $request->thirdPartyPolicy : null;
                 $thirdPartyVehicleRegNo = isset($request->thirdPartyVehicleRegNo) ? $request->thirdPartyVehicleRegNo : null;
+                $assessmentTypeID = $assessment->assessmentTypeID;
                 if ($request->assessmentReviewType == Config::APPROVE) {
                     $approved = Assessment::where(["id" => $request->assessmentID])->update([
                         "assessmentStatusID" => Config::$STATUSES['ASSESSMENT']['PROVISIONAL-APPROVAL']['id'],
@@ -431,12 +432,43 @@ class HeadAssessorController extends Controller
                         $reviewNote = isset($request->report) ? $request->report : '';
                         $role = Config::$ROLES['HEAD-ASSESSOR'];
 
-                        $message = [
-                            'subject' => $claim->claimNo.'_'.$claim->vehicleRegNo.'_'.$this->functions->curlDate(),
-                            'from' => Config::JUBILEE_NO_REPLY_EMAIL,
-                            'to' => $email,
-                            'replyTo' => Config::JUBILEE_NO_REPLY_EMAIL,
-                            'html' => "
+                        if ($assessmentTypeID == Config::ASSESSMENT_TYPES['TOTAL_LOSS']){
+                            $cc_emails = array(User::find(Auth::id())['email'], $email);
+//                            $cc_emails = array(User::find(Auth::id())['email'], 'oumasylvester61@gmail.com');
+                            $message = [
+                                'subject' => $claim->claimNo.'_'.$claim->vehicleRegNo.'_'.$this->functions->curlDate(),
+                                'from' => Config::JUBILEE_NO_REPLY_EMAIL,
+                                'to' => Config::JUBILEE_ALLIANZ_TALK_TO_US_EMAIL,
+//                                'to' => 'sylvesterouma282@gmail.com',
+                                'cc'=>$cc_emails,
+                                'replyTo' => Config::JUBILEE_NO_REPLY_EMAIL,
+                                'html' => "
+                        Hello " . Config::JUBILEE_ALLIANZ_TALK_TO_US_NAME . ", <br>
+
+                        This is in regards to claim number <strong>" . $claimNo . " </strong>, it is deemed a <strong>TOTAL LOSS</strong>. <br> <br>
+
+                        Kindly cancel this policy immediately and flag this vehicle for future policies. <br> <br>
+
+                            <b><i><u>Notes</u></i></b> <br>
+
+                            <i> " . $reviewNote . " </i><br><br>
+
+                        Regards, <br><br>
+
+                        " . $role . ", <br>
+
+                        Claims Department, <br>
+
+                        Jubilee Allianz Insurance Company
+                    ",
+                            ];
+                        }else{
+                            $message = [
+                                'subject' => $claim->claimNo.'_'.$claim->vehicleRegNo.'_'.$this->functions->curlDate(),
+                                'from' => Config::JUBILEE_NO_REPLY_EMAIL,
+                                'to' => $email,
+                                'replyTo' => Config::JUBILEE_NO_REPLY_EMAIL,
+                                'html' => "
                         Hello " . $firstName . ", <br>
 
                         This is in regards to claim number <strong>" . $claimNo . " </strong> <br>
@@ -455,7 +487,9 @@ class HeadAssessorController extends Controller
 
                         Jubilee Allianz Insurance Company
                     ",
-                        ];
+                            ];
+                        }
+
 
                         InfobipEmailHelper::sendEmail($message, $email);
 //                        SMSHelper::sendSMS('Hello ' . $firstName . ', Assessment for claimNo ' . $claimNo . ' has been provisionally approved', $MSISDN);
@@ -501,7 +535,7 @@ class HeadAssessorController extends Controller
                         Jubilee Allianz Insurance Company
                     ",
                                 ];
-                        InfobipEmailHelper::sendEmail($message, $assessmentManager->email);
+                                InfobipEmailHelper::sendEmail($message, $assessmentManager->email);
 //                        SMSHelper::sendSMS('Hello ' . $firstName . ', Assessment for claimNo ' . $claimNo . ' has been provisionally approved', $MSISDN);
 //                        Notification::send($assessor, new ClaimApproved($claim));
 
@@ -514,46 +548,53 @@ class HeadAssessorController extends Controller
 
                         //send the demand letter to third party insurer
 
-                        $assessmentID= $request->assessmentID;
-                        $assessment = Assessment::where(["id"=>$assessmentID])->first();
-                        $claim = Claim::where(["id"=>$assessment->claimID])->with('customer')->first();
-                        $company = Company::where(["id"=>$assessment->companyID])->first();
-                        $adjuster_email = User::where('id', $claim->createdBy)->first()->email;
-                        $cc_emails=array($adjuster_email, Config::JUBILEE_REPLY_EMAIL);
+                        if ($assessment->isSubrogate == 1){
+                            $assessmentID= $request->assessmentID;
+                            $assessment = Assessment::where(["id"=>$assessmentID])->first();
+                            $claim = Claim::where(["id"=>$assessment->claimID])->with('customer')->first();
+                            $company = Company::where(["id"=>$assessment->companyID])->first();
+                            $adjuster_email = User::where('id', $claim->createdBy)->first()->email;
+                            $cc_emails=array($adjuster_email, Config::JUBILEE_REPLY_EMAIL);
+
+
+                            $end_salutation_email = Company::where("name", "=", "JUBILEE ALLIANZ GENERAL INSURANCE (K) LIMITED")->first()->recovery_officer_email;
+                            $end_salutation_first=explode('@', $end_salutation_email)[0];
+                            $first_array=explode(".", $end_salutation_first);
+                            $regards=implode(" ", $first_array);
+                            $pdf = App::make('dompdf.wrapper');
+                            $pdf->loadView('reports.demand-letter', ['assessment'=>$assessment,'claim'=>$claim,'company'=>$company, 'regards'=>$regards]);
+
+                            $pdfName = $assessment['claim']['vehicleRegNo'].'_'.$assessment['claim']['claimNo'];
+                            $pdfName = str_replace("/","_",$pdfName);
+                            $pdfFileName=preg_replace('/\s+/', ' ', $pdfName);
+                            $pdfFileName = str_replace(" ","_",$pdfFileName);
+                            $directoryPath = public_path('reports');
+
+                            if (!File::isDirectory($directoryPath)) {
+                                File::makeDirectory($directoryPath, 0755, true); // Creates the directory recursively with permissions 0755
+                            }
+                            $pdfFilePath = $directoryPath.'/'.$pdfFileName.'.pdf';
+
+
+                            if (File::exists($pdfFilePath)) {
+                                File::delete($pdfFilePath);
+                            }
+                            $pdf->save($pdfFilePath);
+
+                            $flag = false;
 
 
 
-                        $pdf = App::make('dompdf.wrapper');
-                        $pdf->loadView('reports.demand-letter', ['assessment'=>$assessment,'claim'=>$claim,'company'=>$company, 'regards'=>$regards]);
-
-                        $pdfName = $assessment['claim']['vehicleRegNo'].'_'.$assessment['claim']['claimNo'];
-                        $pdfName = str_replace("/","_",$pdfName);
-                        $pdfFileName=preg_replace('/\s+/', ' ', $pdfName);
-                        $pdfFileName = str_replace(" ","_",$pdfFileName);
-                        $pdfFilePath = public_path('reports/'.$pdfFileName.'.pdf');
-                        if (File::exists($pdfFilePath)) {
-                            File::delete($pdfFilePath);
-                        }
-                        $pdf->save($pdfFilePath);
-
-
-                        $flag = false;
-
-                        $end_salutation_email = Company::where("name", "=", "JUBILEE ALLIANZ GENERAL INSURANCE (K) LIMITED")->first()->recovery_officer_email;
-                        $end_salutation_first=explode('@', $end_salutation_email)[0];
-                        $first_array=explode(".", $end_salutation_first);
-                        $regards=implode(" ", $first_array);
-
-                        $message = [
-                            'subject' => "DEMAND LETTER - ".$assessment['claim']['claimNo']."_".$assessment['claim']['vehicleRegNo'],
-                            'from' => Config::JUBILEE_NO_REPLY_EMAIL,
-                            'to' => $company->recovery_officer_email,
+                            $message = [
+                                'subject' => "DEMAND LETTER - ".$assessment['claim']['claimNo']."_".$assessment['claim']['vehicleRegNo'],
+                                'from' => Config::JUBILEE_NO_REPLY_EMAIL,
+                                'to' => $company->recovery_officer_email,
 //                'to' => "sylvesterouma282@gmail.com",
-                            'replyTo' => Config::JUBILEE_NO_REPLY_EMAIL,
-                            'attachment' => $pdfFilePath,
-                            'cc' => $cc_emails,
+                                'replyTo' => Config::JUBILEE_NO_REPLY_EMAIL,
+                                'attachment' => $pdfFilePath,
+                                'cc' => $cc_emails,
 //                'cc' => Auth::user()->email,
-                            'html' => "
+                                'html' => "
 
                         Dear Sirs, <br>
 
@@ -569,11 +610,12 @@ class HeadAssessorController extends Controller
 
                         Jubilee Allianz Insurance Company
                     ",
-                        ];
+                            ];
 
-                        $emailSSent=InfobipEmailHelper::sendEmail($message);
-                        if ($emailSSent){
-                            Assessment::where(["id"=>$assessmentID])->update(["demandLetterDate"=>\Illuminate\Support\Carbon::now(), "subrogationSender"=>Auth::user()->id]);
+                            $emailSSent=InfobipEmailHelper::sendEmail($message);
+                            if ($emailSSent){
+                                Assessment::where(["id"=>$assessmentID])->update(["demandLetterDate"=>\Illuminate\Support\Carbon::now(), "subrogationSender"=>Auth::user()->id]);
+                            }
                         }
 
 
@@ -598,7 +640,8 @@ class HeadAssessorController extends Controller
         } catch (\Exception $e) {
             $response = array(
                 "STATUS_CODE" => Config::GENERIC_ERROR_CODE,
-                "STATUS_MESSAGE" => Config::GENERIC_ERROR_MESSAGE
+                "STATUS_MESSAGE" => Config::GENERIC_ERROR_MESSAGE,
+//                "STATUS_MESSAGE" => $e->getMessage()
             );
             $this->log->motorAssessmentInfoLogger->info("FUNCTION " . __METHOD__ . " " . " LINE " . __LINE__ .
                 "An exception occurred when trying to approve or halt a claim " . $e->getMessage());
